feat: ensure does not run schedulers interval on tests

+++ src/main/config/schedulers.ts
@@ -1,9 +1,11 @@
 import { makeClearUserDatabase } from '../factories/schedulers/clear-user-database-factory'
 
 export default (): void => {
-  const clearUserDatabase = makeClearUserDatabase()
+  if (process.env.NODE_ENV !== 'test') {
+    const clearUserDatabase = makeClearUserDatabase()
 
-  setInterval(async () => {
-    await clearUserDatabase.deleteInactiveUsers()
-  }, 12 * 60 * 60 * 1000)
+    setInterval(async () => {
+      await clearUserDatabase.deleteInactiveUsers()
+    }, 12 * 60 * 60 * 1000)
+  }
 }
